Consolidate tasks and split out features code
