+++ pyroclast/luigi/features_tasks.py
@@ -1,6 +1,30 @@
 import luigi
+import functools
+import itertools
+import json
+import random
 import os.path as osp
-import subprocess
+
+
+class TopTask(luigi.WrapperTask):
+
+    def requires(self):
+        return self.train_tasks()
+
+    def train_tasks(self):
+        arg_dict = {
+            'dataset': ['mnist'],
+            'batch_size': [128],
+            'seed': [8549],
+            'learning_rate': [1e-4],
+            'conv_stack_name': ['attack_net'],
+            'lambd': [0.]
+        }
+        args = [
+            dict(zip(arg_dict.keys(), arg_vals))
+            for arg_vals in itertools.product(*arg_dict.values())
+        ]
+        return [TrainTask(**a) for a in args]
 
 
 class TrainTask(luigi.Task):
